# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drone.Works API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "def-0": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pointer",
          "message"
        ],
        "properties": {
          "pointer": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "title": "ProblemError"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/organizations": {
      "post": {
        "operationId": "createOrganization",
        "summary": "Create an organization",
        "tags": [
          "organizations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "name",
                  "default_timezone",
                  "unit_system"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "default_timezone": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "unit_system": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "metric"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "imperial"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The organization and owner membership were created.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "organization_id",
                    "name",
                    "default_timezone",
                    "unit_system",
                    "role",
                    "pilot_profile_id"
                  ],
                  "properties": {
                    "organization_id": {
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "default_timezone": {
                      "type": "string"
                    },
                    "unit_system": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "metric"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "imperial"
                          ]
                        }
                      ]
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "pilot"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "viewer"
                          ]
                        }
                      ]
                    },
                    "pilot_profile_id": {
                      "anyOf": [
                        {
                          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "The request was denied or could not be accepted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "The request could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/selection": {
      "put": {
        "operationId": "selectOrganization",
        "summary": "Select an organization",
        "tags": [
          "organizations"
        ],
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "organization_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The current canonical membership was selected.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "organization_id",
                    "name",
                    "default_timezone",
                    "unit_system",
                    "role",
                    "pilot_profile_id"
                  ],
                  "properties": {
                    "organization_id": {
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "default_timezone": {
                      "type": "string"
                    },
                    "unit_system": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "metric"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "imperial"
                          ]
                        }
                      ]
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "pilot"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "viewer"
                          ]
                        }
                      ]
                    },
                    "pilot_profile_id": {
                      "anyOf": [
                        {
                          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "The request was denied or could not be accepted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "The request could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}